Split enums from class
